Changed mask var in CBitWriter::WriteBits and CBitReader::ReadBits to uint64_t to account for edge cases where a left shift was exactly 32bits. Changed shifting constants to unsigned long long (ULL) to account for compiler defaulting integers to 32bit

+++ Core/src/BitWriter.cpp
@@ -104,11 +104,11 @@ void CBitWriter::WriteBits(uint32_t data, const int bits)
 		}
 
 		// Create mask
-		unsigned mask;
-		mask = (1 << overflowBits) - 1;
+		uint64_t mask;
+		mask = (1ULL << overflowBits) - 1ULL;
 
 		// From the scratch, grab the bits that overflowed.
-		uint32_t overflow = m_scratch & mask;
+		uint64_t overflow = m_scratch & mask;
 
 		// Shift the scratch "overflowBits" to the right, thus removing the
 		// overflow for safe casting to a 32 bit number
